+++ src/D20Tek.BlazorComponents.Tiles/GlobalUsings.cs
@@ -1,3 +1,5 @@
 global using D20Tek.BlazorComponents.Utilities;
 global using Microsoft.AspNetCore.Components;
 global using Microsoft.AspNetCore.Components.Web;
+global using System;
+global using System.Diagnostics;